Update and rename .gitignore to analog_modulation.Isf
This script is used to analyze the analog modulation on a optoelectronic modulator which is a P-i-N junction in essential. Though the simulation can be done on GUI, scripts can are more flexible as well as efficient if we want to study the effect of a parameter or something else.